+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Supported import and export feeds in JSON format. (#317)
+- Supported choosing auto expand content for each feed. (#306)
 
 ### Changed